feat(config): enhance environment variable support for configuration and protocol parameters
- Added `READ_CONFIG_FROM_ENV` and `READ_PARAMSET_FROM_ENV` options to control configuration source selection.
- Updated `.env.example` to reflect new environment variable options.
- Enhanced documentation in `README.md` to explain configuration methods and source selection logic.
- Refactored CLI to utilize new environment variable logic for loading configurations and protocol parameters.
