+++ .env.example
@@ -1,3 +1,5 @@
+READ_CONFIG_FROM_ENV=1
+
 HOST=127.0.0.1
 PORT=17000
 INDEX=0
@@ -31,8 +33,7 @@ HEADER_CHAIN_PROOF_PATH=../core/tests/data/first_1.bin
 VERIFIER_ENDPOINTS=http://127.0.0.1:17001,http://127.0.0.1:17002,http://127.0.0.1:17003,http://127.0.0.1:17004
 OPERATOR_ENDPOINTS=http://127.0.0.1:17005,http://127.0.0.1:17006
 
-PROTOCOL_PARAMSET=regtest
-
+READ_PARAMSET_FROM_ENV=1
 
 NETWORK=regtest
 NUM_ROUND_TXS=3
@@ -41,7 +42,7 @@ NUM_SIGNED_KICKOFFS=2
 BRIDGE_AMOUNT=1000000000
 KICKOFF_AMOUNT=55000
 OPERATOR_CHALLENGE_AMOUNT=200000000
-COLLATERAL_FUNDING_AMOUNT=200000000 
+COLLATERAL_FUNDING_AMOUNT=200000000
 KICKOFF_BLOCKHASH_COMMIT_LENGTH=40
 WATCHTOWER_CHALLENGE_BYTES=144
 WINTERNITZ_LOG_D=4
