+++ core/src/cli.rs
@@ -3,6 +3,7 @@
 //! This module defines command line interface for binaries. `Clap` is used
 //! for easy generation of help messages and handling arguments.
 
+use crate::config::protocol::ProtocolParamset;
 use crate::config::protocol::ProtocolParamsetName;
 use crate::config::BridgeConfig;
 use crate::errors::BridgeError;
@@ -73,12 +74,52 @@ pub enum ConfigSource {
     File(PathBuf),
     Env,
 }
-
+/// Selects a configuration source for the main config or the protocol paramset.
+///
+/// Configuration can be loaded either from a file specified by a path in the CLI args,
+/// or from environment variables.
+///
+/// Selection logic is as follows:
+///
+/// 1. If the named environment variable (eg. `READ_CONFIG_FROM_ENV`) is not set
+///    or if the named environment variable is set to `0` or `off`, we use the file
+///    path provided in the CLI args (fail if not provided)
+///
+/// 2. If the named environment variable is set to `1` or `on`, we explicitly read from the
+///    environment variable
+///
+/// 3. If the named environment variable is set to an unknown value, we print a
+///    warning and default to environment variables
+///
+/// # Examples
+///
+/// ```bash
+/// # Load config from a file and protocol params from a file
+/// READ_CONFIG_FROM_ENV=0 READ_PARAMSET_FROM_ENV=0 clementine-core verifier --config-file /path/to/config.toml --protocol-params-file /path/to/protocol-params.toml
+///
+/// # or
+/// # define all config variables in the environment
+/// export CONFIG_ONE=1
+/// export PARAM_ONE=1
+/// # and source from environment variables
+/// READ_CONFIG_FROM_ENV=1 READ_PARAMSET_FROM_ENV=1 clementine-core verifier
+///
+/// # or
+/// # source paramset from environment variables but use config from a file
+/// export PARAM_ONE=1
+/// export PARAM_TWO=1
+/// READ_CONFIG_FROM_ENV=0 READ_PARAMSET_FROM_ENV=1 clementine-core verifier --config-file /path/to/config.toml
+///
+/// # WRONG usage (will use environment variables for both config and paramset)
+/// export CONFIG_ONE=1
+/// export PARAM_ONE=1
+/// READ_CONFIG_FROM_ENV=1 READ_PARAMSET_FROM_ENV=1 clementine-core --config-file /path/to/config.toml --protocol-params-file /path/to/protocol-params.toml
+/// ```
 pub fn get_config_source(
-    env_name: &'static str,
+    read_from_env_name: &'static str,
     provided_arg: Option<PathBuf>,
 ) -> Result<ConfigSource, BridgeError> {
-    Ok(match std::env::var(env_name) {
+    Ok(match std::env::var(read_from_env_name) {
         Err(_) => ConfigSource::File(provided_arg.ok_or(BridgeError::ConfigError(
             "No file path or environment variable provided for config file.".to_string(),
         ))?),
@@ -87,11 +128,11 @@ pub fn get_config_source(
         )?),
         Ok(str) => {
             if str != "1" || str != "on" {
-                tracing::warn!("Unknown value for {env_name}: {str}. Expected 1/0/off/on. Defaulting to environment variables.");
+                tracing::warn!("Unknown value for {read_from_env_name}: {str}. Expected 1/0/off/on. Defaulting to environment variables.");
             }
 
             if provided_arg.is_some() {
-                tracing::warn!("File path provided in CLI arguments while {env_name} is set to 1. Ignoring provided file path and reading from environment variables.");
+                tracing::warn!("File path provided in CLI arguments while {read_from_env_name} is set to 1. Ignoring provided file path and reading from environment variables.");
             }
 
             ConfigSource::Env
@@ -106,8 +147,10 @@ pub fn get_config_source(
 ///
 /// 1. Get CLI arguments
 /// 2. Initialize logger
-/// 3. Get configuration file, either from environment variables or
+/// 3. Get configuration, either from environment variables or
 ///    configuration file
+/// 4. Get protocol parameters, either from environment variables or
+///    protocol parameters file
 ///
 /// # Returns
 ///
@@ -115,7 +158,7 @@ pub fn get_config_source(
 ///
 /// - [`BridgeConfig`] from CLI argument
 /// - [`Args`] from CLI options
-pub fn get_configuration_from_cli() -> (BridgeConfig, Args) {
+pub fn get_cli_config() -> (BridgeConfig, Args) {
     let args = match parse_args() {
         Ok(args) => args,
         Err(e) => {
@@ -139,37 +182,18 @@ pub fn get_configuration_from_cli() -> (BridgeConfig, Args) {
         }
     };
 
-    let protocol_params_source =
-        get_config_source("READ_PARAMSET_FROM_ENV", args.protocol_params_file.clone())
-            .wrap_err("Failed to determine source for protocol parameters.");
-
-    let paramset = match protocol_params_source {
-        Err(e) => {
-            tracing::error!("Failed to determine source for protocol parameters: {e:?}");
-            delayed_panic!("Failed to determine source for protocol parameters: {e:?}");
-        }
-        Ok(ConfigSource::File(path)) => {
-            // TODO: Temporary solution, make it better
-            std::env::set_var("PROTOCOL_CONFIG_PATH", path.to_str().unwrap_or_default());
-            ProtocolParamsetName::File
-        }
-        Ok(ConfigSource::Env) => ProtocolParamsetName::Env,
-    };
-
     let config_source = get_config_source("READ_CONFIG_FROM_ENV", args.config_file.clone())
         .wrap_err("Failed to determine source for configuration.");
 
     let (mut config, args) = match config_source {
         Err(e) => {
-            tracing::error!("Failed to determine source for configuration: {e:?}");
             delayed_panic!("Failed to determine source for configuration: {e:?}");
         }
         Ok(ConfigSource::File(config_file)) => {
             // Read from configuration file ONLY
-            let config = match read_config_from(config_file) {
+            let config = match BridgeConfig::try_parse_file(config_file) {
                 Ok(config) => config,
                 Err(e) => {
-                    tracing::error!("Can't read configuration from file: {e:?}");
                     delayed_panic!("Can't read configuration from file: {e:?}");
                 }
             };
@@ -184,19 +208,12 @@ pub fn get_configuration_from_cli() -> (BridgeConfig, Args) {
                     let e = e.into_eyre();
                     match e.root_cause().downcast_ref::<BridgeError>() {
                         Some(BridgeError::EnvVarNotSet(e, field)) => {
-                            tracing::error!(
-                                "Missing environment variable {field} in environment config mode. ({e:?})."
-                            );
                             delayed_panic!("Missing environment variable {field} in environment config mode. ({e:?}).");
                         }
                         Some(BridgeError::EnvVarMalformed(e, field)) => {
-                            tracing::error!("Malformed environment variable {field} in environment config mode. ({e:?}).");
                             delayed_panic!("Malformed environment variable {field} in environment config mode. ({e:?}).");
                         }
                         _ => {
-                            tracing::error!(
-                            "Error occurred while reading environment variables for config: {e:?}"
-                            );
                             delayed_panic!("Error occurred while reading environment variables for config: {e:?}. ({e:?}).");
                         }
                     }
@@ -205,6 +222,31 @@ pub fn get_configuration_from_cli() -> (BridgeConfig, Args) {
         }
     };
 
+    let protocol_params_source =
+        get_config_source("READ_PARAMSET_FROM_ENV", args.protocol_params_file.clone());
+
+    // Leaks memory to get a static reference to the paramset
+    // This is needed to reduce copies of the protocol paramset when passing it around.
+    // This is fine, since this will only run once in the lifetime of the program.
+    let paramset: &'static ProtocolParamset = Box::leak(Box::new(match protocol_params_source {
+        Err(e) => {
+            delayed_panic!("Failed to determine source for protocol parameters: {e:?}");
+        }
+        Ok(ConfigSource::File(path)) => match ProtocolParamset::from_toml_file(path.as_path()) {
+            Ok(paramset) => paramset,
+            Err(e) => {
+                delayed_panic!("Failed to read protocol parameters from file: {e:?}");
+            }
+        },
+        Ok(ConfigSource::Env) => match ProtocolParamset::from_env() {
+            Ok(paramset) => paramset,
+            Err(e) => {
+                delayed_panic!("Failed to read protocol parameters from environment: {e:?}");
+            }
+        },
+    }));
+
+    // The default will be REGTEST_PARAMSET and is overridden from the selected source above.
     config.protocol_paramset = paramset;
 
     (config, args)
